saki-kernel-sdk/pyproject.toml
[project]
name = "saki-kernel-sdk"
version = "0.1.0"
description = "Saki kernel runtime SDK (IPC/manifest/error handling)"
requires-python = ">=3.10"
dependencies = [
  "pyzmq>=26.0.0",
  "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/saki_kernel_sdk"]
